refactor: #54 - Alert창 생성 부 위치 변경, Alertable 변경사항 롤백
