+++ Targets/Core/Sources/Extensions/Notification+.swift
@@ -10,6 +10,5 @@ import Foundation
 
 public extension Notification.Name {
   static let muteAllPlayers = Notification.Name("muteAllPlayers")
-  static let reSignIn = Notification.Name("reSignIn")
   static let logout = Notification.Name("logout")
 }
